Split the package up into multiple entrypoints
This helps group related components together nicely.
