+++ src/Tab/index.ts
@@ -1,0 +1,10 @@
+export { default as createTabGroup } from "src/Tab/createTabGroup";
+export { default as TabItem } from "src/Tab/TabItem";
+export { default as TabList } from "src/Tab/TabList";
+export { default as TabPanel } from "src/Tab/TabPanel";
+export { default as TabProvider, useTabContext } from "src/Tab/TabProvider";
+
+export type { TabItemProps } from "src/Tab/TabItem";
+export type { TabListProps } from "src/Tab/TabList";
+export type { TabPanelProps } from "src/Tab/TabPanel";
+export type { TabProviderProps, TabContextValue } from "src/Tab/TabProvider";
