--- src/Tabs/index.ts
@@ -1,10 +1,0 @@
-export { default as createTabGroup } from "src/Tabs/createTabGroup";
-export { default as TabItem } from "src/Tabs/TabItem";
-export { default as TabList } from "src/Tabs/TabList";
-export { default as TabPanel } from "src/Tabs/TabPanel";
-export { default as TabProvider, useTabContext } from "src/Tabs/TabProvider";
-
-export type { TabItemProps } from "src/Tabs/TabItem";
-export type { TabListProps } from "src/Tabs/TabList";
-export type { TabPanelProps } from "src/Tabs/TabPanel";
-export type { TabProviderProps, TabContextValue } from "src/Tabs/TabProvider";
