+++ src/audio/index.ts
@@ -1,0 +1,4 @@
+export { default as AudioControls } from "src/audio/AudioControls";
+export { default as AudioProvider, useAudioContext } from "src/audio/AudioProvider";
+
+export type { AudioContextValue, AudioProviderProps, TrackData } from "src/audio/AudioProvider";
