+++ src/form/createFormHook.ts
@@ -3,8 +3,8 @@ import type { ComponentType } from "react";
 import { omitProperties } from "@alextheman/utility";
 import { createFormHook as createTanstackFormHook } from "@tanstack/react-form";
 
-import { SubmitButton, TextField } from "src/v7/components";
-import { fieldContext, formContext } from "src/v7/hooks/formHooks";
+import { SubmitButton, TextField } from "src/form";
+import { fieldContext, formContext } from "src/form/formHooks";
 
 export interface AlexFieldComponents {
   TextField: typeof TextField;
