+++ src/snackbar/index.ts
@@ -1,0 +1,3 @@
+export { default as SnackbarProvider, useSnackbar } from "src/snackbar/SnackbarProvider";
+
+export type { SnackbarProviderProps } from "src/snackbar/SnackbarProvider";
