+++ src/theme/ThemeToggle.tsx
@@ -2,7 +2,7 @@ import Tooltip from "@mui/material/Tooltip";
 import { MdOutlineDarkMode, MdOutlineLightMode } from "react-icons/md";
 
 import SwitchWithIcons from "src/root/components/SwitchWithIcons";
-import { useTheme } from "src/root/providers";
+import { useTheme } from "src/theme/ThemeProvider";
 
 /** A toggle to switch between dark mode and light mode. Must be used in a `ThemeProvider`. */
 function ThemeToggle() {
