+++ src/theme/index.ts
@@ -1,0 +1,3 @@
+export { default as ThemeProvider, useTheme } from "src/theme/ThemeProvider";
+
+export type { ThemeProviderProps, ThemeContextValue } from "src/theme/ThemeProvider";
