--- src/v7/components/DropdownMenu/index.tsx
@@ -1,17 +1,0 @@
-export { default as DropdownMenu } from "src/v7/components/DropdownMenu/DropdownMenu";
-export { default as DropdownMenuItem } from "src/v7/components/DropdownMenu/DropdownMenuItem";
-export {
-  default as DropdownMenuProvider,
-  useDropdownMenu,
-} from "src/v7/components/DropdownMenu/DropdownMenuProvider";
-export { default as DropdownMenuTrigger } from "src/v7/components/DropdownMenu/DropdownMenuTrigger";
-export { default as DropdownMenuWrapper } from "src/v7/components/DropdownMenu/DropdownMenuWrapper";
-
-export type { DropdownMenuProps } from "src/v7/components/DropdownMenu/DropdownMenu";
-export type { DropdownMenuItemProps } from "src/v7/components/DropdownMenu/DropdownMenuItem";
-export type {
-  DropdownMenuProviderProps,
-  DropdownMenuContextValue,
-} from "src/v7/components/DropdownMenu/DropdownMenuProvider";
-export type { DropdownMenuTriggerProps } from "src/v7/components/DropdownMenu/DropdownMenuTrigger";
-export type { DropdownMenuWrapperProps } from "src/v7/components/DropdownMenu/DropdownMenuWrapper";
