--- src/v7/components/index.ts
@@ -1,11 +1,0 @@
-export { default as NavigationBottom } from "src/v7/components/NavigationBottom";
-export { default as NavigationDrawer } from "src/v7/components/NavigationDrawer";
-export { default as SubmitButton } from "src/v7/components/SubmitButton";
-export { default as TextField } from "src/v7/components/TextField";
-
-export * from "src/v7/components/DropdownMenu";
-export * from "src/v7/components/routing";
-
-export type { NavigationBottomProps, NavItemBottom } from "src/v7/components/NavigationBottom";
-export type { NavigationDrawerProps, NavMenuItem } from "src/v7/components/NavigationDrawer";
-export type { SubmitButtonProps } from "src/v7/components/SubmitButton";
