--- src/v7/components/routing/index.ts
@@ -1,10 +1,0 @@
-export { default as ErrorPage } from "src/v7/components/routing/ErrorPage";
-export { default as InternalLink } from "src/v7/components/routing/InternalLink";
-export { default as MemoryRouter } from "src/v7/components/routing/MemoryRouter";
-export { default as Router } from "src/v7/components/routing/Router";
-export { default as Switch } from "src/v7/components/routing/Switch";
-export { default as useAbsoluteLocation } from "src/v7/components/routing/useAbsoluteLocation";
-
-export type { ErrorPageProps } from "src/v7/components/routing/ErrorPage";
-export type { InternalLinkProps } from "src/v7/components/routing/InternalLink";
-export type { SwitchProps } from "src/v7/components/routing/Switch";
