--- src/v7/hooks/index.ts
@@ -1,3 +1,0 @@
-export { default as createFormHook } from "src/v7/hooks/createFormHook";
-
-export * from "src/v7/hooks/formHooks";
