--- src/v7/index.ts
@@ -1,2 +1,0 @@
-export * from "src/v7/components";
-export * from "src/v7/hooks";
