+++ stories/NavigationDrawer.stories.tsx
@@ -2,8 +2,10 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 
 import Button from "@mui/material/Button";
 
-import { Page, ThemeProvider, ThemeToggle } from "src/root";
-import { MemoryRouter, NavigationDrawer } from "src/v7";
+import { Page } from "src/root";
+import { MemoryRouter, NavigationDrawer } from "src/routing";
+import { ThemeProvider } from "src/theme";
+import ThemeToggle from "src/theme/ThemeToggle";
 
 const meta: Meta<typeof NavigationDrawer> = {
   component: NavigationDrawer,
