+++ tsdown.config.ts
@@ -14,16 +14,40 @@ const baseConfig: UserConfig = {
 
 const config: Array<UserConfig> = [
   {
-    entry: ["src/root/index.ts"],
-    outDir: "dist",
+    entry: ["src/audio/index.ts"],
+    outDir: "dist/audio",
+  },
+  {
+    entry: ["src/DropdownMenu/index.ts"],
+    outDir: "dist/DropdownMenu",
+  },
+  {
+    entry: ["src/form/index.ts"],
+    outDir: "dist/form",
   },
   {
     entry: ["src/QueryBoundary/index.ts"],
     outDir: "dist/QueryBoundary",
   },
   {
-    entry: ["src/Tabs/index.ts"],
-    outDir: "dist/Tabs",
+    entry: ["src/root/index.ts"],
+    outDir: "dist",
+  },
+  {
+    entry: ["src/routing/index.ts"],
+    outDir: "dist/routing",
+  },
+  {
+    entry: ["src/snackbar/index.ts"],
+    outDir: "dist/snackbar",
+  },
+  {
+    entry: ["src/Tab/index.ts"],
+    outDir: "dist/Tab",
+  },
+  {
+    entry: ["src/theme/index.ts"],
+    outDir: "dist/theme",
   },
 ].map(({ entry, outDir }) => {
   return {
